requirements.txt
lightning==2.4
torch==2.0.0
torch-spatiotemporal==0.9.5
pandas==2.2.3
numpy==1.26.4
scipy==1.14.1
dtw-python==1.5.3
einops==0.8.0
torch-geometric==2.6.0
torch-scatter==2.1.2
torch-sparse==0.6.18